requirements.txt
python-telegram-bot
zalgotext
